requirements.txt
asgiref==3.4.1
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet @ file:///C:/ci/chardet_1607706937985/work
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
conda==4.10.3
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262410900/work
cryptography==3.4.8
Django==3.2.6
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-crum==0.7.9
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-solo==1.1.5
ecdsa==0.17.0
fido2==0.9.1
idna==3.2
intelhex==2.3.0
menuinst==1.4.16
Pillow==8.3.1
psycopg2==2.9.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyserial==3.5
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-decouple==3.4
pytz==2021.1
pyusb==1.2.1
pywin32==228
requests==2.26.0
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.1
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
